+++ src/main/java/com/ngt/jopenmetaverse/shared/protocol/ObjectUpdateCompressedPacket.java
@@ -29,7 +29,7 @@ package com.ngt.jopenmetaverse.shared.protocol;
                 try
                 {
                     RegionHandle = Utils.bytesToULong(bytes, i[0]); i[0] += 8;
-                    TimeDilation = (ushort)Utils.bytesToUInt16(bytes, i[0]); i[0] += 2;
+                    TimeDilation = (int)Utils.bytesToUInt16(bytes, i[0]); i[0] += 2;
                 }
                 catch (Exception e)
                 {
